% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altimetry-antarctica.R
\name{altimetry_antarctica_files}
\alias{altimetry_antarctica_files}
\title{Antarctic Polar Altimetry files}
\usage{
altimetry_antarctica_files()
}
\value{

}
\description{
Sea Level Anomaly measured by Altimetry and derived variables
}
\details{
sla, formal_error, U, V
}
\examples{
\dontrun{
  aaf <- altimetry_antarctica_files()
}
}
