% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-altimetry-files.R
\name{altimetry_currents_polar_files}
\alias{altimetry_currents_polar_files}
\title{Derived altimetry products}
\usage{
altimetry_currents_polar_files(hemisphere = "south")
}
\arguments{
\item{hemisphere}{south only for now}
}
\description{
A polar-transformed copy of the 'u' and 'v' components of surface currents from [altimetry_daily_files].
}
\details{
Only available for the southern hemisphere.
}
\examples{
altimetry_currents_polar_files()
}
