% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-argo-files.R
\name{argo_files}
\alias{argo_files}
\title{Argo files.}
\usage{
argo_files(type = c("prof", "meta", "traj", "tech", "Mprof"), dac = NULL)
}
\arguments{
\item{type}{file type, i.e. prof, meta, traj, tech}

\item{dac}{data acqusition centre e.g.* "aoml", "bodc", "coriolis", "csio", "csiro", "incois", all returned if not specified}
}
\description{
ARGO
}
