% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfafiles.R
\name{bom_tmax_daily_files}
\alias{bom_tmax_daily_files}
\title{BOM tmax daily fils}
\usage{
bom_tmax_daily_files()
}
\value{
data frame of file names
}
\description{
Data is daily but arranged in monthly files.
}
