% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-cafe-modis.R
\name{cafe_monthly_files}
\alias{cafe_monthly_files}
\title{'Cafe' MODIS files}
\usage{
cafe_monthly_files()
}
\value{
data frame of file names and date, 'date', 'fullname'
}
\description{
'Cafe' MODIS files
}
\examples{
\dontrun{
  cafe_monthly_files()
}
}
