% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmpwind-files.R
\name{ccmp}
\alias{ccmp}
\alias{ccmp_6hourly_files}
\title{Derived surface files from Remote Sensing Systems (RSS)}
\usage{
ccmp_6hourly_files()
}
\value{
tibble data frame of file names, with columns `fullname` and `date`
}
\description{
RSS CCMP_RT V2.1 derived surface winds (Level 3.0), variables 'uwnd', 'vwnd', 'nobs'
'u-wind vector component at 10 meters', 'v-wind vector component at 10 meters', and
'number of observations used to derive wind vector components' from [Remote Sensing Systems](http://www.remss.com/).
}
\details{
Each file contains four time steps at six hourly intervals aligned to the
file base date.
}
\examples{
\dontrun{
  ccmp_6hourly_files()
}
}
\references{
"Mears et al., Journal of Geophysical Research: Oceans,124,
  6997-7010, 2019, Hoffman et al., Journal of Atmospheric and Oceanic
  Technology, 2013; Atlas et al., BAMS, 2011; Atlas et al., BAMS, 1996". "Available for
  public use with proper citation".
}
