% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice-cersat-files.R
\name{cersat_daily_files}
\alias{cersat_daily_files}
\title{CERSAT daily  sea-ice concentration}
\usage{
cersat_daily_files()
}
\value{
tibble data frame of file names
}
\description{
Sea ice concentration files at 12.5 km resolution, southern hemisphere.
}
\examples{
\dontrun{
  cersat_daily_files()
}
}
