% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip5.R
\name{cmip5_files}
\alias{cmip5_files}
\title{CMIP5 files}
\usage{
cmip5_files()
}
\value{
data frame of file names
}
\description{
Currently in dev-test mode, using a very small subset until we get specifications.
}
