% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-fsle-files.R
\name{fsle_files}
\alias{fsle_files}
\title{Backward-in-time Finite-Size Lyapunov Exponents.}
\usage{
fsle_files()
}
\description{
FSLE - MAPS OF FINITE SIZE LYAPUNOV EXPONENTS AND ORIENTATIONS OF THE ASSOCIATED EIGENVECTORS
}
\details{
These are daily files.
}
\references{
[https://www.aviso.altimetry.fr/en/data/products/value-added-products/fsle-finite-size-lyapunov-exponents.html]( Finite-Size Lyapunov Exponents)
}
