% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-geoid-files.R
\name{geoid_files}
\alias{geoid_files}
\title{Earth Gravitation Model files}
\usage{
geoid_files(all = FALSE, ...)
}
\arguments{
\item{all}{return all files, or just the core grid files for GDAL? './w001001.adf'}
}
\value{
data frame of file paths
}
\description{
Global 2.5 Minute Geoid Undulations, a
}
\details{
Each file is an ESRI GRID raster data set of 2.5-minute geoid undulation values covering a 45 x 45 degree area.
Each raster file has a 2.5-minute cell size and is a subset of the global 2.5 x 2.5-minute grid of pre-computed
geoid undulation point values found on the EGM2008-WGS 84 Version web page. This ESRI GRID format represents a
continuous surface of geoid undulation values where each 2.5-minute raster cell derives its value from the original
pre-computed geoid undulation point value located at the SW corner of each cell.
}
\examples{
geoid_files()
}
