% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadfiles.R
\name{ghrsst}
\alias{ghrsst}
\alias{ghrsst_daily_files}
\title{GHRSST files}
\usage{
ghrsst_daily_files()
}
\value{
tibble data frame of file names
}
\description{
The Group for High Resolution Sea Surface Temperature (GHRSST) files.
}
\examples{
oisst_daily_files()
}
