% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst-files.R
\name{ghrsst_files}
\alias{ghrsst_files}
\alias{ghrsst_daily_files}
\alias{ghrsst_daily_files_netcdf}
\title{GHRSST files}
\usage{
ghrsst_daily_files()

ghrsst_daily_files_netcdf()
}
\value{
tibble data frame of file names
}
\description{
The Group for High Resolution Sea Surface Temperature (GHRSST) files.
}
\examples{
\dontrun{
  ghrsst_daily_files()
}
}
