% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncep2-files.R
\name{ncep2_files}
\alias{ncep2_files}
\alias{ncep2_uwnd_6hr_files}
\alias{ncep2_files}
\alias{ncep2_vwnd_6hr_files}
\title{NCEP2 wind files}
\usage{
ncep2_uwnd_6hr_files()

ncep2_vwnd_6hr_files()
}
\value{
tibble data frame of file names
}
\description{
NCEP2 six-hourly reanalysis2 gaussian grid
}
\examples{
ncep2_uwnd_6hr_files()
ncep2_vwnd_6hr_files()
}
