% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_colour_l3.R
\name{ocean_colour_files}
\alias{ocean_colour_files}
\title{AQUA_MODIS.20121210_20121217.L3m.8D.PAR.par.4km.nc'}
\usage{
ocean_colour_files(
  temporal = c("", "DAY", "8D", "MO", "R32"),
  spatial = c("", "9km", "4km"),
  product = c("^[[:alpha:]]{2}", "AQUA_MODIS", "SEASTAR_SEAWIFS", "SNPP_VIIRS"),
  varname = c("", "CHL", "PAR", "POC", "KD490"),
  type = c("", "L3m", "L3b"),
  bz2.rm = TRUE,
  ext = c("", "nc", "main"),
  ...,
  silent = FALSE
)
}
\description{
AQUA_MODIS.20121210_20121217.L3m.8D.PAR.par.4km.nc'
}
