% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc-files.R, R/raad-ice-amsr-files.R,
%   R/raad-sst-files.R
\name{oisst}
\alias{oisst}
\alias{nsidc_south_monthly_files}
\alias{oisst}
\alias{amsr_daily_files}
\alias{oisst}
\alias{oisst_daily_files}
\alias{oisst}
\alias{oisst_monthly_files}
\title{NSIDC daily and monthly sea-ice concentration}
\usage{
nsidc_south_monthly_files()

amsr_daily_files()

oisst_daily_files()

oisst_monthly_files()
}
\value{
tibble data frame of file names

tibble data frame of file names

tibble data frame of file names
}
\description{
Sea ice concentration files.

Sea ice concentration files at 6.25 km resolution, southern hemisphere.

Optimally Interpolated Sea Surface Temperature
}
\examples{
nsidc_south_monthly_files()
nsidc_north_monthly_files()
nsidc_monthly_files()
amsr_daily_files()
oisst_daily_files()
}
