% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadfiles.R
\name{oisst_daily_files}
\alias{oisst_daily_files}
\title{OISST files}
\usage{
oisst_daily_files()
}
\value{
tibble data frame of file names
}
\description{
Optimally Interpolated Sea Surface Temperature
}
\examples{
oisst_daily_files()
}
