% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarview.R
\name{polarview}
\alias{polarview}
\alias{polarview_files}
\title{Polarview files}
\usage{
polarview_files(type = c("jpeg", "tarball"))
}
\arguments{
\item{type}{jpeg or tarball}
}
\value{
tibble data frame of file names, with columns `fullname` and `date`
}
\description{
Imagery from www.polarview.aq
}
\details{
The JPEGs are simple images, the GeoTIFFs are 16-bit integers (haven't explored further)
}
\examples{
\dontrun{
  files <- polarview_files()
  tiffiles <- polarview_files(type = "tarball")
}
}
