% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_raad_data_roots}
\alias{get_raad_data_roots}
\alias{get_raad_filenames}
\alias{set_raad_data_roots}
\alias{raad_filedb_path}
\alias{set_raad_filenames}
\title{Raadfile administration tools}
\usage{
get_raad_data_roots()

get_raad_filenames()

set_raad_data_roots(..., replace_existing = TRUE,
  use_known_candidates = FALSE)

raad_filedb_path(...)

set_raad_filenames(clobber = FALSE)
}
\arguments{
\item{...}{input file paths to set}

\item{replace_existing}{replace existing paths, defaults to TRUE}

\item{use_known_candidates}{apply internal logic for known candidates (for internal use at raad-hq), defaults to FALSE}

\item{clobber}{by default do not ignore existing file cache, set to TRUE to ignore and set}
}
\description{
Administration tools for managing a data library.
}
\details{
`get_raad_data_roots` returns the current list of visible root directories

`get_raad_filenames` returns the entire list of all files found in visible root directories

`set_raad_filenames` runs the system to update the file listing and refresh it

`set_raad_data_roots` set data root paths, default use is to apply the strings input as the exclusive set

`raad_filedb_path` return the path to the database of file names for each input directory
}
