% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadfiles-package.R
\docType{package}
\name{raadfiles-package}
\alias{raadfiles-package}
\alias{raadfiles}
\title{Manage files used provided by bowerbird/raadtools.}
\description{
Tools for managing collections of files for the 'raad' family.
}
\seealso{
raadfiles-admin
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{michael.sumner@aad.gov.au}

Other contributors:
\itemize{
  \item Ben Raymond [contributor]
  \item Kimberlee Baldry (\href{https://orcid.org/0000-0003-3286-8624}{ORCID}) [contributor]
}

}
\keyword{internal}
