% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_par.R
\name{read_par}
\alias{read_par}
\title{Data frame of file names and dates.}
\usage{
read_par(
  date,
  time.resolution = "8daily",
  xlim = NULL,
  lon180 = FALSE,
  nobsonly = FALSE,
  latest = TRUE,
  returnfiles = FALSE,
  inputfiles = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Data frame of file names and dates.
}
