% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadfiles-package.R
\docType{data}
\name{rema_tiles_polygons}
\alias{rema_tiles_polygons}
\title{Rema tile polygons}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1516 rows and 3 columns.
}
\usage{
rema_tiles_polygons
}
\description{
Rema tile polygons
}
\keyword{datasets}
