% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-seapodym-files.R
\name{seapodym_weekly_files}
\alias{seapodym_weekly_files}
\title{SEAPODYM model files}
\usage{
seapodym_weekly_files()
}
\description{
Global ocean low and mid trophic levels biomass hindcast
}
\references{
\url{http://www.cls.fr}, \url{http://www.seapodym.eu}
}
