% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_raad_time_stamp}
\alias{set_raad_time_stamp}
\alias{get_raad_time_stamp}
\title{Set a time stamp on a data frame}
\usage{
set_raad_time_stamp(x)

get_raad_time_stamp()
}
\arguments{
\item{x}{data frame}
}
\value{
x with attribute "raad_time_stamp" set
}
\description{
Used by `set_raad_filenames` when uploading the cache to memory.
}
