% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap-files.R
\name{smap}
\alias{smap}
\alias{smap_8day_files}
\alias{salt}
\title{SMAP ocean surface salinity files}
\usage{
smap_8day_files()
}
\value{
tibble data frame of file names, with columns `fullname` and `date`
}
\description{
Remote Sensing Systems SMAP Level 3 Sea Surface Salinity Standard Mapped Image 8day running
}
\examples{
\dontrun{
  smap_daily_files()
}
}
