% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography-files.R
\name{topography}
\alias{topography}
\alias{gebco21_files}
\alias{gebco19_files}
\alias{gebco14_files}
\alias{gebco08_files}
\alias{ramp_files}
\alias{ibcso_files}
\alias{ibcso_background_files}
\alias{ibcso_bed_files}
\alias{ibcso_digital_chart_files}
\alias{ibcso_sid_files}
\alias{cryosat2_files}
\alias{etopo1_files}
\alias{etopo2_files}
\alias{lakesuperior_files}
\alias{kerguelen_files}
\alias{george_v_terre_adelie_1000m_files}
\alias{george_v_terre_adelie_500m_files}
\alias{george_v_terre_adelie_250m_files}
\alias{george_v_terre_adelie_100m_files}
\alias{smith_sandwell_files}
\alias{smith_sandwell_unpolished_files}
\alias{smith_sandwell_lon180_files}
\alias{smith_sandwell_unpolished_lon180_files}
\alias{macquarie100m_57S_files}
\alias{macquarie100m_58S_files}
\title{Topographic data files}
\usage{
gebco21_files(all = FALSE, ...)

gebco19_files(all = FALSE, ...)

gebco14_files(all = FALSE, ...)

gebco08_files(all = FALSE, ...)

ramp_files(all = FALSE, ...)

ibcso_files(all = FALSE, ...)

ibcso_background_files(all = FALSE, ...)

ibcso_bed_files(all = FALSE, ...)

ibcso_digital_chart_files(all = FALSE, ...)

ibcso_sid_files(all = FALSE, ...)

cryosat2_files(all = FALSE, ...)

etopo1_files(all = FALSE, ...)

etopo2_files(all = FALSE, ...)

lakesuperior_files(all = FALSE, ...)

kerguelen_files(all = FALSE, ...)

george_v_terre_adelie_1000m_files(all = FALSE, ...)

george_v_terre_adelie_500m_files(all = FALSE, ...)

george_v_terre_adelie_250m_files(all = FALSE, ...)

george_v_terre_adelie_100m_files(all = FALSE, ...)

smith_sandwell_files(all = FALSE, ...)

smith_sandwell_unpolished_files(all = FALSE, ...)

smith_sandwell_lon180_files(all = FALSE, ...)

smith_sandwell_unpolished_lon180_files(all = FALSE, ...)

macquarie100m_57S_files(all = FALSE, ...)

macquarie100m_58S_files(all = FALSE, ...)
}
\arguments{
\item{all}{return a larger set of files (for exploratory use only)}

\item{...}{reserved}
}
\value{
data frame of `file` and `fullname` columns
}
\description{
Obtain file names for various topographic data.
}
\details{
Each function exists to match a specific data set, but the optional `all` argument may be used to
easily discover a broader set of files that ship with the data, or that represent older versions, documentation
and other metadata files.

There's no single format, there are GeoTIFFs, ArcInfo binary, ERStorage, NetCDF, NetCDF GMT, (Geo) PDF, and some VRT
wrappers for handling raw binary files.
}
\section{GEBCO General Bathymetric Chart of the Oceans}{

Versions 2008, 2014, 2019, 2021.
}

\section{IBCSO International Bathymetric Chart of the Southern Ocean}{

'is' ('is_PS71' tif, or grd), 'background_hq', 'bed' ('bed_PS71'), 'digital_chart', 'sid' ('sid_PS71')
}

\section{ETOPO}{

Etopo1 and Etopo2, Lake Superior
}

\section{Smith and Sandwell}{

Polished and unpolished, version 18.1  replaces 18.
}

\section{Cryosat2}{

Cryosat2
}

\section{Australian Antarctic}{

George V Terre Adelie, Kerguelen, Macquarie 100m
}

\examples{
\dontrun{
  gebco21_files()
}

}
