% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-world-ocean-atlas-files.R
\name{WOA}
\alias{WOA}
\alias{woa13_files}
\alias{woa09_files}
\alias{woa09_daily_files}
\title{World Ocean Atlas  products}
\usage{
woa13_files()

woa09_files()

woa09_daily_files()
}
\value{
tibble data frame of file names
}
\description{
WOA find files
}
\details{
Current returns all NetCDF files, without any date information, there's a mix of variables month/year climatologies.
}
\examples{
woa13_daily_files()
}
