% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-altimetry-files.R
\name{altimetry}
\alias{altimetry}
\alias{altimetry_daily_files}
\title{Altimetry and so forth}
\usage{
altimetry_daily_files()
}
\value{
tibble data frame of file names
}
\description{
From Copernicus
}
\examples{
altimetry_daily_files()
}
