% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-altimetry-files.R,
%   R/raad-world-ocean-atlas-files.R
\name{altimetry}
\alias{altimetry}
\alias{altimetry_daily_files}
\alias{altimetry}
\alias{woa09_daily_files}
\title{Altimetry products}
\usage{
altimetry_daily_files()

woa09_daily_files()
}
\value{
tibble data frame of file names

tibble data frame of file names
}
\description{
From Copernicus

WOA09 incomplete function
}
\details{
Current returns all NetCDF files, without any date.
}
\examples{
altimetry_daily_files()
woa09_daily_files()
}
