% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-ice-amsr-files.R
\name{amsr_daily_files}
\alias{amsr_daily_files}
\title{AMSR daily  sea-ice concentration}
\usage{
amsr_daily_files()
}
\value{
tibble data frame of file names
}
\description{
Sea ice concentration files at 6.25 km resolution, southern hemisphere.
}
\examples{
amsr_daily_files()
}
