% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo-files.R
\name{argo_files}
\alias{argo_files}
\title{Argo files.}
\usage{
argo_files(type = c("prof", "meta", "traj", "tech", "Mprof"), dac = NULL)
}
\arguments{
\item{type}{file type, i.e. prof, meta, traj, tech}

\item{dac}{data acqusition centre e.g.* "aoml", "bodc", "coriolis", "csio", "csiro", "incois", all returned if not specified}
}
\description{
ARGO, by default we have 'profiles', alternatively 'meta' for type
}
\details{
(No traj, tech, or Mprof at this time of writing 2022-11-21)
}
