% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice-leads.R
\name{leads}
\alias{leads}
\alias{iceclim_south_leadsfiles}
\alias{iceclim_north_leadsfiles}
\title{Files containing relative lead frequencies for the Arctic and Antarctic}
\usage{
iceclim_south_leadsfiles(all = FALSE)

iceclim_north_leadsfiles(all = FALSE)
}
\arguments{
\item{all}{return all files, or just the core grid files (*.nc)?}
}
\description{
Average Lead-Frequency for the polar oceans for winter months November-April 2002/03-2018/19 based on daily lead
composites as derived from MOD/MYD-29 IST 5 min granules.
}
\examples{
iceclim_south_leadsfiles()
iceclim_north_leadsfiles()
}
\references{
F. Reiser, S. Willmes, G. Heinemann (2020): A new algorithm for daily sea ice lead identification in the Arctic and
 Antarctic winter from thermal-infrared satellite imagery.
}
