% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc-files.R, R/nsidc_cdr-files.R
\name{nsidc}
\alias{nsidc}
\alias{nsidc_south_monthly_files}
\alias{nsidc_north_monthly_files}
\alias{nsidc_monthly_files}
\alias{nsidc_south_daily_files}
\alias{nsidc_north_daily_files}
\alias{nsidc_daily_files}
\alias{nsidc_daily_files_v2}
\alias{nsidc_monthly_files_v2}
\alias{nsidc_cdr_daily_files}
\alias{nsidc_cdr_monthly_files}
\title{NSIDC daily and monthly sea-ice concentration}
\usage{
nsidc_south_monthly_files()

nsidc_north_monthly_files()

nsidc_monthly_files()

nsidc_south_daily_files()

nsidc_north_daily_files()

nsidc_daily_files()

nsidc_daily_files_v2(extra_pattern = NULL)

nsidc_monthly_files_v2(extra_pattern = NULL)

nsidc_south_monthly_files()

nsidc_north_monthly_files()

nsidc_monthly_files()

nsidc_south_daily_files()

nsidc_north_daily_files()

nsidc_daily_files()

nsidc_cdr_daily_files(extra_pattern = NULL)

nsidc_cdr_monthly_files(extra_pattern = NULL)
}
\arguments{
\item{extra_pattern}{argument for restricted string matching}
}
\value{
tibble data frame of file names

tibble data frame of file names
}
\description{
Sea ice concentration files.

Sea ice concentration files.
}
\examples{
\dontrun{
  nsidc_south_monthly_files()
  nsidc_north_monthly_files()
  nsidc_monthly_files()
  nsidc_south_daily_files()
  nsidc_north_daily_files()
  nsidc_daily_files()
}
\dontrun{
  nsidc_south_monthly_files()
  nsidc_north_monthly_files()
  nsidc_monthly_files()
  nsidc_south_daily_files()
  nsidc_north_daily_files()
  nsidc_daily_files()
}
}
