% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc-files.R
\name{nsidc}
\alias{nsidc}
\alias{nsidc_south_monthly_files}
\alias{nsidc_north_monthly_files}
\alias{nsidc_monthly_files}
\alias{nsidc_south_daily_files}
\alias{nsidc_north_daily_files}
\alias{nsidc_daily_files}
\title{NSIDC daily and monthly sea-ice concentration}
\usage{
nsidc_south_monthly_files()

nsidc_north_monthly_files()

nsidc_monthly_files()

nsidc_south_daily_files()

nsidc_north_daily_files()

nsidc_daily_files()
}
\value{
tibble data frame of file names
}
\description{
Sea ice concentration files.
}
\examples{
nsidc_south_monthly_files()
nsidc_north_monthly_files()
nsidc_monthly_files()
nsidc_south_daily_files()
nsidc_north_daily_files()
nsidc_daily_files()
}
