% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-sst-files.R
\name{oisst}
\alias{oisst}
\alias{oisst_daily_files}
\alias{oisst_monthly_files}
\title{OISST v2 files}
\usage{
oisst_daily_files()

oisst_monthly_files()
}
\value{
tibble data frame of file names, with columns `fullname` and `date`
}
\description{
Optimally Interpolated Sea Surface Temperature, from \url{https://www.ncei.noaa.gov/}. These files contain four
variables `sst`, `anom`, `err` and `ice` for sea surface temperature, sst anomaly, sst error and sea ice concentration on
a regular global longitude latitude grid, with dimensions 1440x720 grid (0.25 degree spatial resolution).
}
\details{
At the time of writing (2021-01-18) the files are accessible at
\url{https://www.ncei.noaa.gov/data/sea-surface-temperature-optimum-interpolation/v2.1/access/avhrr/}. See the
[blueant](https://github.com/AustralianAntarcticDivision/blueant) package for a convenient way to obtain this data set
named "NOAA OI 1/4 Degree Daily SST AVHRR".

These files can be accessed individually `raster` package function `raster` or as multiple layers with `brick` or `raster::stack`. Use
the `varname` argument to choose one of the four variables.

To obtain full NetCDF header metadata use 'ncdf4::open.nc(file)' or 'RNetCDF::print.nc(RNetCDF::open.nc(file))' to see
the equivalent of 'ncdump -h' output.

Optimally Interpolated version 2 SST moved from 'eclipse.ncdc.noaa.gov', to 'www.ncei.noaa.gov' at the end of 2017. Version 2 was superseded by version 2.1 during 2020.
}
\examples{
\dontrun{
  oisst_daily_files()
}
}
