% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc-files.R, R/raad-sst-files.R
\name{oisst}
\alias{oisst}
\alias{nsidc_south_monthly_files}
\alias{oisst}
\alias{oisst_daily_files}
\alias{oisst}
\alias{oisst_monthly_files}
\title{NSIDC monthly sea-ice concentration}
\usage{
nsidc_south_monthly_files()

oisst_daily_files()

oisst_monthly_files()
}
\value{
tibble data frame of file names

tibble data frame of file names
}
\description{
Sea ice concentration files.

Optimally Interpolated Sea Surface Temperature
}
\examples{
oisst_daily_files()
}
