% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-sst-files.R
\name{oisst}
\alias{oisst}
\alias{oisst_daily_files}
\alias{oisst}
\alias{oisst_monthly_files}
\title{OISST files}
\usage{
oisst_daily_files()

oisst_monthly_files()
}
\value{
tibble data frame of file names
}
\description{
Optimally Interpolated Sea Surface Temperature
}
\examples{
oisst_daily_files()
}
