% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raadfiles-package.R
\docType{package}
\name{raadfiles-package}
\alias{raadfiles}
\alias{raadfiles-package}
\title{raadfiles: File Database Management for 'raadtools'}
\description{
Tools for managing collections of files for the 'raad' family.
}
\seealso{
raadfiles-admin
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{michael.sumner@aad.gov.au}

Other contributors:
\itemize{
  \item Ben Raymond [contributor]
}

}
\keyword{internal}
