% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rema-files.R
\name{rema_8m_files}
\alias{rema_8m_files}
\alias{rema_tile_files}
\alias{rema_8m_files}
\alias{rema_100m_files}
\alias{rema_8m_files}
\alias{rema_200m_files}
\alias{rema_8m_files}
\alias{rema_1km_files}
\alias{rema_8m_files}
\alias{rema_100m_files}
\alias{rema_200m_files}
\alias{rema_1km_files}
\alias{rema_tile_files}
\title{Files for The Reference Elevation Model of Antarctica (REMA)}
\usage{
rema_tile_files(all = FALSE, ...)

rema_100m_files(filled = TRUE, ...)

rema_200m_files()

rema_1km_files(filled = TRUE, ...)

rema_8m_files(...)
}
\arguments{
\item{...}{}

\item{filled}{return 'filled' variant if available}
}
\value{
data frame of file names
}
\description{
Return files for various products from REMA Release 1
}
\details{
'rema_8m_files' returns the base level 8 GeoTIFF files, there are 1516 files at 8m resolution.
}
\examples{
rema_8m_files()
rema_100m_files(filled = TRUE)
}
\references{
\url{https://www.pgc.umn.edu/tag/rema/}
}
