% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raad-bsose.R
\name{sose_monthly_files}
\alias{sose_monthly_files}
\title{Southern Ocean State Estimate files}
\usage{
sose_monthly_files(iteration = "")
}
\arguments{
\item{iteration}{default is '' which finds latest available, see details}
}
\value{
data frame of file names and date, 'date', 'fullname'
}
\description{
Files from Ocean State Estimation at Scripps for the Southern Ocean 'SOSE'.
}
\details{
Iteration provided is latest available, otherwise this argument will be used to match with file names.
}
\examples{
sose_monthly_files()
}
\references{
http://sose.ucsd.edu/
}
